% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ebs_sppMaster}
\alias{ebs_sppMaster}
\title{spp_master for eastern bering sea}
\format{data.table}
\usage{
ebs_sppMaster
}
\description{
spp_master is an object from trawlDiversity
}
\keyword{datasets}

