% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDist.R
\name{geoDist}
\alias{geoDist}
\title{geographic distance}
\usage{
geoDist(x, y)
}
\arguments{
\item{x, y}{length-2 (or 2-column object coercable to matrix) where the first element (or column) is longtitude, second is latitude}
}
\value{
returns the distance between \code{x} and \code{y}. If \code{x} and \code{y} have multiple rows, then returns a vector of the geographic distances between corresponding rows
}
\description{
geographic distance between two sets of points
}

