% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mapOwin}
\alias{mapOwin}
\title{regional outlines}
\format{owin}
\usage{
mapOwin
}
\description{
mapOwin is an object from trawlDiversity; is a \code{spatstat::owin} object. Includes outlines for all regions in trawlData, including eastern bering sea
}
\keyword{datasets}

