% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot map}
\usage{
plot_map(x, y, marks, map_owin)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{marks}{value to be plotted}

\item{owin}{object for outline}
}
\value{
invisibly returns \code{NULL}
}
\description{
plots a smoothed map of a value as a semi-transparent background w/ outline of sampled area and landmass map
}

